Split test_storage_iceberg
